python/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
psycopg2-binary
itsdangerous
pymongo
httpx
python-dotenv
